chatbot/requirements.txt
Flask==3.0.2
openai==1.14.3
python-dotenv==1.0.1
